# docs/swagger/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/rol": {
      "post": {
        "operationId": "RolController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolRegisterDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Rol"
        ]
      },
      "get": {
        "operationId": "RolController_findAllPagination",
        "parameters": [
          {
            "name": "orderProperty",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "permissions",
                "typeRol"
              ]
            }
          },
          {
            "name": "permission",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SUPER",
                "USER",
                "USER_PROFILE",
                "USER_ADD",
                "USER_READ",
                "USER_READ_ONE",
                "USER_UPDATE",
                "USER_DELETE",
                "ROL",
                "ROL_ADD",
                "ROL_READ",
                "ROL_READ_ONE",
                "ROL_UPDATE",
                "ROL_DELETE",
                "PATHOLOGY",
                "PATHOLOGY_ADD",
                "PATHOLOGY_READ",
                "PATHOLOGY_READ_ONE",
                "PATHOLOGY_UPDATE",
                "PATHOLOGY_DELETE",
                "MEDICAMENT",
                "MEDICAMENT_ADD",
                "MEDICAMENT_READ",
                "MEDICAMENT_READ_ONE",
                "MEDICAMENT_UPDATE",
                "MEDICAMENT_DELETE",
                "MEDICATIONREQUEST",
                "MEDICATIONREQUEST_ADD",
                "MEDICATIONREQUEST_READ",
                "MEDICATIONREQUEST_READ_ONE",
                "MEDICATIONREQUEST_UPDATE",
                "MEDICATIONREQUEST_ASSIGN",
                "MEDICATIONREQUEST_DELIVERY_DATA",
                "MEDICATIONREQUEST_DELETE",
                "MEDICATIONREQUEST_PHARMACY",
                "MEDICATIONREQUEST_VALIDATE_PATIENT",
                "PHARMACY",
                "PHARMACY_ADD",
                "PHARMACY_READ",
                "PHARMACY_READ_ONE",
                "PHARMACY_UPDATE",
                "PHARMACY_DELETE",
                "PATIENTS",
                "PATIENTS_ADD",
                "PATIENTS_READ",
                "PATIENTS_READ_ONE",
                "PATIENTS_UPDATE",
                "PATIENTS_DELETE",
                "DELIVERYDATA",
                "DELIVERYDATA_ADD",
                "DELIVERYDATA_READ",
                "DELIVERYDATA_READ_ONE",
                "DELIVERYDATA_UPDATE",
                "DELIVERYDATA_DELETE",
                "INVENTORY",
                "INVENTORY_ADD",
                "INVENTORY_READ",
                "INVENTORY_READ_ONE",
                "INVENTORY_UPDATE",
                "INVENTORY_DELETE",
                "ASSIGNATION",
                "ASSIGNATION_ADD",
                "ASSIGNATION_READ",
                "ASSIGNATION_READ_ONE",
                "ASSIGNATION_UPDATE",
                "ASSIGNATION_DELETE",
                "ASSIGNATION_PDF",
                "INCIDENTREPORT",
                "INCIDENTREPORT_ADD",
                "INCIDENTREPORT_READ",
                "INCIDENTREPORT_READ_ONE",
                "INCIDENTREPORT_UPDATE",
                "INCIDENTREPORT_DELETE",
                "INCIDENTREPORT_UPDATE_STATUS",
                "INCIDENTREPORT_PDF",
                "AUDIT",
                "AUDIT_READ",
                "AUDIT_DELETE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Rol"
        ]
      },
      "patch": {
        "operationId": "RolController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolUpdateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Rol"
        ]
      }
    },
    "/api/rol/one/{uid}": {
      "get": {
        "operationId": "RolController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Rol"
        ]
      }
    },
    "/api/rol/per": {
      "get": {
        "operationId": "RolController_findPer",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Rol"
        ]
      }
    },
    "/api/rol/all": {
      "get": {
        "operationId": "RolController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Rol"
        ]
      }
    },
    "/api/rol/delete/{uid}": {
      "delete": {
        "operationId": "RolController_delete",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Rol"
        ]
      }
    },
...
    "/api/audit": {
      "get": {
        "operationId": "AuditController_findAllPagination",
        "parameters": [
          {
            "name": "orderProperty",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ci",
                "first_name",
                "first_surname",
                "ip",
                "userAgent",
                "userPlatform"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Audit"
        ]
      }
    },
    "/api/audit/delete/{uid}": {
      "delete": {
        "operationId": "AuditController_delete",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Audit"
        ]
      }
    },
...
    "/api/image": {
      "post": {
        "operationId": "HandelImageController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "HandelImage"
        ]
      }
    },
    "/api/image/{name}": {
      "get": {
        "operationId": "HandelImageController_getImg",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "HandelImage"
        ]
      }
    },
...
  },
  "info": {
    "title": "Documentación API",
    "description": "Esta es la documentación de la API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RolRegisterDTO": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "permissions": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string",
              "enum": [
                "SUPER",
                "USER",
                "USER_PROFILE",
                "USER_ADD",
                "USER_READ",
                "USER_READ_ONE",
                "USER_UPDATE",
                "USER_DELETE",
                "ROL",
                "ROL_ADD",
                "ROL_READ",
                "ROL_READ_ONE",
                "ROL_UPDATE",
                "ROL_DELETE",
                "PATHOLOGY",
                "PATHOLOGY_ADD",
                "PATHOLOGY_READ",
                "PATHOLOGY_READ_ONE",
                "PATHOLOGY_UPDATE",
                "PATHOLOGY_DELETE",
                "MEDICAMENT",
                "MEDICAMENT_ADD",
                "MEDICAMENT_READ",
                "MEDICAMENT_READ_ONE",
                "MEDICAMENT_UPDATE",
                "MEDICAMENT_DELETE",
                "MEDICATIONREQUEST",
                "MEDICATIONREQUEST_ADD",
                "MEDICATIONREQUEST_READ",
                "MEDICATIONREQUEST_READ_ONE",
                "MEDICATIONREQUEST_UPDATE",
                "MEDICATIONREQUEST_ASSIGN",
                "MEDICATIONREQUEST_DELIVERY_DATA",
                "MEDICATIONREQUEST_DELETE",
                "MEDICATIONREQUEST_PHARMACY",
                "MEDICATIONREQUEST_VALIDATE_PATIENT",
                "PHARMACY",
                "PHARMACY_ADD",
                "PHARMACY_READ",
                "PHARMACY_READ_ONE",
                "PHARMACY_UPDATE",
                "PHARMACY_DELETE",
                "PATIENTS",
                "PATIENTS_ADD",
                "PATIENTS_READ",
                "PATIENTS_READ_ONE",
                "PATIENTS_UPDATE",
                "PATIENTS_DELETE",
                "DELIVERYDATA",
                "DELIVERYDATA_ADD",
                "DELIVERYDATA_READ",
                "DELIVERYDATA_READ_ONE",
                "DELIVERYDATA_UPDATE",
                "DELIVERYDATA_DELETE",
                "INVENTORY",
                "INVENTORY_ADD",
                "INVENTORY_READ",
                "INVENTORY_READ_ONE",
                "INVENTORY_UPDATE",
                "INVENTORY_DELETE",
                "ASSIGNATION",
                "ASSIGNATION_ADD",
                "ASSIGNATION_READ",
                "ASSIGNATION_READ_ONE",
                "ASSIGNATION_UPDATE",
                "ASSIGNATION_DELETE",
                "ASSIGNATION_PDF",
                "INCIDENTREPORT",
                "INCIDENTREPORT_ADD",
                "INCIDENTREPORT_READ",
                "INCIDENTREPORT_READ_ONE",
                "INCIDENTREPORT_UPDATE",
                "INCIDENTREPORT_DELETE",
                "INCIDENTREPORT_UPDATE_STATUS",
                "INCIDENTREPORT_PDF",
                "AUDIT",
                "AUDIT_READ",
                "AUDIT_DELETE"
              ]
            }
          },
          "typeRol": {
            "type": "string",
            "enum": [
              "admin",
              "user"
            ]
          }
        },
        "required": [
          "uid",
          "name",
          "description",
          "permissions",
          "typeRol"
        ]
      },
      "Role": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "typeRol": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "boolean"
          },
          "users": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "uid",
          "name",
          "description",
          "typeRol",
          "permissions",
          "status",
          "users"
        ]
      },
...
      "Pharmacy": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "state": {
            "enum": [
              "Amazonas",
              "Anzoátegui",
              "Apure",
              "Aragua",
              "Barinas",
              "Bolívar",
              "Carabobo",
              "Cojedes",
              "Delta Amacuro",
              "Distrito Capital",
              "Falcón",
              "Guarico",
              "Lara",
              "Mérida",
              "Miranda",
              "Monagas",
              "Nueva Esparta",
              "Portuguesa",
              "Sucre",
              "Táchira",
              "Trujillo",
              "Vargas",
              "Yaracuy",
              "Zulia"
            ],
            "type": "string"
          },
          "municipality": {
            "enum": [
              "Alto Orinoco",
              "Atabapo",
              "Atures",
              "Autana",
              "Manapiare",
              "Maroa",
              "Río Negro",
              "Anaco",
              "Aragua",
              "Carvajal",
              "Diego Bautista Urbaneja",
              "Freites",
              "Guanipa",
              "Guanta",
              "McGregor",
              "Peñalver",
              "San Juan de Capistrano",
              "Santa Ana",
              "Achaguas",
              "Biruaca",
              "Muñoz",
              "Pedro Camejo",
              "San Fernando",
              "Camatagua",
              "Francisco Linares Alcántara",
              "José Ángel Lamas",
              "José Félix Ribas",
              "José Rafael Revenga",
              "Mario Briceño Iragorry",
              "Ocumare de la Costa de Oro",
              "San Casimiro",
              "San Sebastián",
              "Santiago Mariño",
              "Santos Michelena",
              "Alberto Arvelo Torrealba",
              "Antonio José de Sucre",
              "Barinas",
              "Cruz Paredes",
              "Obispos",
              "Pedraza",
              "Rojas",
              "Sosa",
              "Angostura",
              "Caroní",
              "El Callao",
              "Gran Sabana",
              "Heres",
              "Píar",
              "Sifontes",
              "Padre Pedro Chien",
              "Bejuma",
              "Carlos Arvelo",
              "Diego Ibarra",
              "Guacara",
              "Mora",
              "Los Guayos",
              "Montalbán",
              "Naguanagua",
              "Puerto Cabello",
              "San Diego",
              "San Joaquín",
              "Valencia",
              "Anzoátegui",
              "Tinaquillo",
              "Girardot",
              "Lima Blanco",
              "Pao de San Juan Bautista",
              "Ricaurte",
              "Rómulo Gallegos",
              "San Carlos",
              "Tinaco",
              "Antonio Díaz",
              "Casacoima",
              "Pedernales",
              "Tucupita",
              "Buchivacoa",
              "Carirubana",
              "Colina",
              "Dabajuro",
              "Democracia",
              "Falcón",
              "Federación",
              "Jacura",
              "Los Taques",
              "Manaure",
              "Mauroa",
              "Monseñor Iturriza",
              "Palmasola",
              "Petit",
              "Píritu",
              "Silva",
              "Tocópero",
              "Unión",
              "Urumaco",
              "Camaguán",
              "Chaguaramas",
              "El Socorro",
              "Infante",
              "Las Mercedes",
              "Mellado",
              "Monagas",
              "Ortiz",
              "Ribas",
              "Roscio",
              "San Gerónimo de Guayabal",
              "San José de Guaribe",
              "Santa María de Ipire",
              "Zaraza",
              "Crespo",
              "Iribarren",
              "Jiménez",
              "Morán",
              "Palavecino",
              "Simón Planas",
              "Torres",
              "Alberto Adriani",
              "Antonio Pinto Salinas",
              "Aricagua",
              "Arzobispo Chacón",
              "Campo Elías",
              "Caracciolo Parra Olmedo",
              "Cardenal Quintero",
              "Guaraque",
              "Julio César Salas",
              "Justo Briceño",
              "Obispo Ramos de Lora",
              "Padre Noguera",
              "Pueblo Llano",
              "Rangel",
              "Rivas Dávila",
              "Santos Marquina",
              "Tovar",
              "Tulio Febres Cordero",
              "Zea",
              "Acevedo",
              "Baruta",
              "Brión",
              "Buroz",
              "Carrizal",
              "Chacao",
              "Cristóbal Rojas",
              "El Hatillo",
              "Guaicaipuro",
              "Lander",
              "Los Salias",
              "Paz Castillo",
              "Pedro Gual",
              "Plaza",
              "Zamora",
              "Acosta",
              "Aguasay",
              "Caripe",
              "Cedeño",
              "Ezequiel Zamora",
              "Maturín",
              "Piar",
              "Punceres",
              "Santa Bárbara",
              "Sotillo",
              "Uracoa",
              "Antolín del Campo",
              "Díaz",
              "García",
              "Gómez",
              "Maneiro",
              "Marcano",
              "Península de Macanao",
              "Tubores",
              "Villalba",
              "Agua Blanca",
              "Araure",
              "Esteller",
              "Guanare",
              "Guanarito",
              "Monseñor José Vicente de Unda",
              "Ospino",
              "Papelón",
              "San Genaro de Boconoíto",
              "San Rafael de Onoto",
              "Santa Rosalía",
              "Turén",
              "Andrés Eloy Blanco",
              "Andrés Mata",
              "Arismendi",
              "Benítez",
              "Bermúdez",
              "Cajigal",
              "Cruz Salmerón Acosta",
              "Mariño",
              "Mejía",
              "Montes",
              "Ribero",
              "Valdez",
              "Antonio Rómulo Costa",
              "Ayacucho",
              "Cárdenas",
              "Córdoba",
              "Fernández Feo",
              "Francisco de Miranda",
              "García de Hevia",
              "Guásimos",
              "Jáuregui",
              "José María Vargas",
              "Junín",
              "Libertad",
              "Lobatera",
              "Michelena",
              "Panamericano",
              "Pedro María Ureña",
              "Rafael Urdaneta",
              "Samuel Darío Maldonado",
              "San Cristóbal",
              "Seboruco",
              "Simón Rodríguez",
              "Torbes",
              "Uribante",
              "San Judas Tadeo",
              "Andrés Bello",
              "Boconó",
              "Candelaria",
              "Carache",
              "Escuque",
              "José Felipe Márquez Cañizalez",
              "Juan Vicente Campos Elías",
              "La Ceiba",
              "Monte Carmelo",
              "Motatán",
              "Pampán",
              "Pampanito",
              "Rafael Rangel",
              "San Rafael de Carvajal",
              "Trujillo",
              "Urdaneta",
              "Valera",
              "Vargas",
              "Arístides Bastidas",
              "Bolívar",
              "Bruzual",
              "Cocorote",
              "Independencia",
              "José Antonio Páez",
              "La Trinidad",
              "Manuel Monge",
              "Nirgua",
              "Peña",
              "San Felipe",
              "Urachiche",
              "Veroes",
              "Almirante Padilla",
              "Baralt",
              "Cabimas",
              "Catatumbo",
              "Colón",
              "Francisco Javier Pulgar",
              "Páez",
              "Jesús Enrique Lossada",
              "Jesús María Semprún",
              "La Cañada de Urdaneta",
              "Lagunillas",
              "Machiques de Perijá",
              "Mara",
              "Maracaibo",
              "Miranda",
              "Rosario de Perijá",
              "San Francisco",
              "Santa Rita",
              "Simón Bolívar",
              "Sucre",
              "Valmore Rodríguez",
              "Libertador"
            ],
            "type": "string"
          },
          "status": {
            "type": "boolean"
          },
          "uidUser": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "medicationRequest": {
            "$ref": "#/components/schemas/MedicationRequest"
          },
          "deliveryData": {
            "$ref": "#/components/schemas/DeliveryData"
          }
        },
        "required": [
          "uid",
          "name",
          "address",
          "state",
          "municipality",
          "status",
          "uidUser",
          "user",
          "medicationRequest",
          "deliveryData"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string"
          },
          "v_e": {
            "enum": [
              "v",
              "e"
            ],
            "type": "string"
          },
          "ci": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "middle_name": {
            "type": "string"
          },
          "first_surname": {
            "type": "string"
          },
          "last_surname": {
            "type": "string"
          },
          "sex": {
            "enum": [
              "f",
              "m"
            ],
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "status": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "activatedAccount": {
            "type": "boolean"
          },
          "attemptCount": {
            "type": "number"
          },
          "dataOfAttempt": {
            "type": "string"
          },
          "uidRol": {
            "type": "string"
          },
          "rol": {
            "$ref": "#/components/schemas/Role"
          },
          "uidPharmacy": {
            "type": "string"
          },
          "pharmacyAS": {
            "$ref": "#/components/schemas/Pharmacy"
          },
          "patients": {
            "$ref": "#/components/schemas/Patients"
          },
          "pharmacy": {
            "$ref": "#/components/schemas/Pharmacy"
          },
          "medicationRequest": {
            "$ref": "#/components/schemas/MedicationRequest"
          },
          "pathology": {
            "$ref": "#/components/schemas/Pathology"
          },
          "medicament": {
            "$ref": "#/components/schemas/Medicament"
          }
        },
        "required": [
          "uid",
          "v_e",
          "ci",
          "first_name",
          "middle_name",
          "first_surname",
          "last_surname",
          "sex",
          "phone",
          "email",
          "password",
          "status",
          "code",
          "activatedAccount",
          "attemptCount",
          "dataOfAttempt",
          "uidRol",
          "rol",
          "uidPharmacy",
          "pharmacyAS",
          "patients",
          "pharmacy",
          "medicationRequest",
          "pathology",
          "medicament"
        ]
      },
      "RolUpdateDTO": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "minLength": 3,
            "maxLength": 255
          },
          "permissions": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "SUPER",
                "USER",
                "USER_PROFILE",
                "USER_ADD",
                "USER_READ",
                "USER_READ_ONE",
                "USER_UPDATE",
                "USER_DELETE",
                "ROL",
                "ROL_ADD",
                "ROL_READ",
                "ROL_READ_ONE",
                "ROL_UPDATE",
                "ROL_DELETE",
                "PATHOLOGY",
                "PATHOLOGY_ADD",
                "PATHOLOGY_READ",
                "PATHOLOGY_READ_ONE",
                "PATHOLOGY_UPDATE",
                "PATHOLOGY_DELETE",
                "MEDICAMENT",
                "MEDICAMENT_ADD",
                "MEDICAMENT_READ",
                "MEDICAMENT_READ_ONE",
                "MEDICAMENT_UPDATE",
                "MEDICAMENT_DELETE",
                "MEDICATIONREQUEST",
                "MEDICATIONREQUEST_ADD",
                "MEDICATIONREQUEST_READ",
                "MEDICATIONREQUEST_READ_ONE",
                "MEDICATIONREQUEST_UPDATE",
                "MEDICATIONREQUEST_ASSIGN",
                "MEDICATIONREQUEST_DELIVERY_DATA",
                "MEDICATIONREQUEST_DELETE",
                "MEDICATIONREQUEST_PHARMACY",
                "MEDICATIONREQUEST_VALIDATE_PATIENT",
                "PHARMACY",
                "PHARMACY_ADD",
                "PHARMACY_READ",
                "PHARMACY_READ_ONE",
                "PHARMACY_UPDATE",
                "PHARMACY_DELETE",
                "PATIENTS",
                "PATIENTS_ADD",
                "PATIENTS_READ",
                "PATIENTS_READ_ONE",
                "PATIENTS_UPDATE",
                "PATIENTS_DELETE",
                "DELIVERYDATA",
                "DELIVERYDATA_ADD",
                "DELIVERYDATA_READ",
                "DELIVERYDATA_READ_ONE",
                "DELIVERYDATA_UPDATE",
                "DELIVERYDATA_DELETE",
                "INVENTORY",
                "INVENTORY_ADD",
                "INVENTORY_READ",
                "INVENTORY_READ_ONE",
                "INVENTORY_UPDATE",
                "INVENTORY_DELETE",
                "ASSIGNATION",
                "ASSIGNATION_ADD",
                "ASSIGNATION_READ",
                "ASSIGNATION_READ_ONE",
                "ASSIGNATION_UPDATE",
                "ASSIGNATION_DELETE",
                "ASSIGNATION_PDF",
                "INCIDENTREPORT",
                "INCIDENTREPORT_ADD",
                "INCIDENTREPORT_READ",
                "INCIDENTREPORT_READ_ONE",
                "INCIDENTREPORT_UPDATE",
                "INCIDENTREPORT_DELETE",
                "INCIDENTREPORT_UPDATE_STATUS",
                "INCIDENTREPORT_PDF",
                "AUDIT",
                "AUDIT_READ",
                "AUDIT_DELETE"
              ]
            }
          },
          "typeRol": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string"
          },
          "status": {
            "type": "boolean"
          }
        },
        "required": [
          "uid",
          "name",
          "description",
          "permissions",
          "typeRol",
          "status"
        ]
      },
...
    }
  }
}